+++ elements/gmf_ai_audio/examples/common/gmf_setup/esp_gmf_setup_peripheral.c
@@ -7,7 +7,6 @@
 #include "esp_log.h"
 #include "driver/sdmmc_host.h"
 #include "vfs_fat_internal.h"
-#include "esp_wifi.h"
 #include "esp_event.h"
 #include "nvs_flash.h"
 #include "esp_gmf_oal_mem.h"
@@ -25,9 +24,11 @@
 #include "esp_codec_dev_defaults.h"
 #endif  /* USE_ESP_GMF_ESP_CODEC_DEV_IO */
 
-#define WIFI_CONNECTED_BIT     BIT0
-#define WIFI_FAIL_BIT          BIT1
-#define WIFI_RECONNECT_RETRIES 30
+typedef enum {
+    I2S_CREATE_MODE_TX_ONLY   = 0,
+    I2S_CREATE_MODE_RX_ONLY   = 1,
+    I2S_CREATE_MODE_TX_AND_RX = 2,
+} i2s_create_mode_t;
 
 #ifdef USE_ESP_GMF_ESP_CODEC_DEV_IO
 i2s_chan_handle_t            rx_handle   = NULL;
@@ -43,22 +44,8 @@ const audio_codec_if_t      *out_codec_if = NULL;
 const audio_codec_gpio_if_t *gpio_if = NULL;
 #endif  /* USE_ESP_GMF_ESP_CODEC_DEV_IO */
 
-#ifndef CONFIG_EXAMPLE_ESP_GMF_WIFI
-#define CONFIG_EXAMPLE_ESP_GMF_WIFI_SSID "****"
-#define CONFIG_EXAMPLE_ESP_GMF_WIFI_PASS "****"
-#endif  /* CONFIG_EXAMPLE_ESP_GMF_WIFI */
-
-typedef enum {
-    I2S_CREATE_MODE_TX_ONLY   = 0,
-    I2S_CREATE_MODE_RX_ONLY   = 1,
-    I2S_CREATE_MODE_TX_AND_RX = 2,
-} i2s_create_mode_t;
-
-static const char        *TAG = "SETUP_PERIPH";
-static esp_netif_t       *s_esp_gmf_setup_sta_netif;
-static EventGroupHandle_t s_wifi_event_group;
-static int                s_retry_num = 0;
-i2c_master_bus_handle_t   i2c_handle  = NULL;
+static const char      *TAG        = "SETUP_PERIPH";
+i2c_master_bus_handle_t i2c_handle = NULL;
 
 #ifdef USE_ESP_GMF_ESP_CODEC_DEV_IO
 static esp_err_t setup_periph_i2s_tx_init(esp_gmf_setup_periph_aud_info *aud_info)
@@ -340,92 +327,6 @@ void esp_gmf_setup_periph_sdmmc(void **out_card)
 void esp_gmf_teardown_periph_sdmmc(void *card)
 {
     esp_vfs_fat_sdcard_unmount("/sdcard", card);
-}
-
-static void event_handler(void *arg, esp_event_base_t event_base,
-                          int32_t event_id, void *event_data)
-{
-    if (event_base == WIFI_EVENT && event_id == WIFI_EVENT_STA_START) {
-        esp_wifi_connect();
-    } else if (event_base == WIFI_EVENT && event_id == WIFI_EVENT_STA_DISCONNECTED) {
-        if (s_retry_num < WIFI_RECONNECT_RETRIES) {
-            esp_wifi_connect();
-            s_retry_num++;
-            ESP_LOGI(TAG, "Retry to connect to the AP");
-        } else {
-            xEventGroupSetBits(s_wifi_event_group, WIFI_FAIL_BIT);
-        }
-        ESP_LOGI(TAG, "connect to the AP fail");
-    } else if (event_base == IP_EVENT && event_id == IP_EVENT_STA_GOT_IP) {
-        ip_event_got_ip_t *event = (ip_event_got_ip_t *)event_data;
-        ESP_LOGI(TAG, "got ip:" IPSTR, IP2STR(&event->ip_info.ip));
-        s_retry_num = 0;
-        xEventGroupSetBits(s_wifi_event_group, WIFI_CONNECTED_BIT);
-    }
-}
-
-void esp_gmf_setup_periph_wifi(void)
-{
-    esp_err_t ret = nvs_flash_init();
-    if (ret == ESP_ERR_NVS_NO_FREE_PAGES || ret == ESP_ERR_NVS_NEW_VERSION_FOUND) {
-        ESP_ERROR_CHECK(nvs_flash_erase());
-        ret = nvs_flash_init();
-    }
-    ESP_ERROR_CHECK(ret);
-    s_wifi_event_group = xEventGroupCreate();
-    ESP_ERROR_CHECK(esp_netif_init());
-    ESP_ERROR_CHECK(esp_event_loop_create_default());
-    s_esp_gmf_setup_sta_netif = esp_netif_create_default_wifi_sta();
-    wifi_init_config_t cfg = WIFI_INIT_CONFIG_DEFAULT();
-    ESP_ERROR_CHECK(esp_wifi_init(&cfg));
-    esp_event_handler_instance_t instance_any_id;
-    esp_event_handler_instance_t instance_got_ip;
-    ESP_ERROR_CHECK(esp_event_handler_instance_register(WIFI_EVENT,
-                                                        ESP_EVENT_ANY_ID,
-                                                        &event_handler,
-                                                        NULL,
-                                                        &instance_any_id));
-    ESP_ERROR_CHECK(esp_event_handler_instance_register(IP_EVENT,
-                                                        IP_EVENT_STA_GOT_IP,
-                                                        &event_handler,
-                                                        NULL,
-                                                        &instance_got_ip));
-    wifi_config_t wifi_config = {
-        .sta = {
-            .ssid = CONFIG_EXAMPLE_ESP_GMF_WIFI_SSID,
-            .password = CONFIG_EXAMPLE_ESP_GMF_WIFI_PASS,
-        },
-    };
-    ESP_ERROR_CHECK(esp_wifi_set_mode(WIFI_MODE_STA));
-    ESP_ERROR_CHECK(esp_wifi_set_config(WIFI_IF_STA, &wifi_config));
-    ESP_ERROR_CHECK(esp_wifi_start());
-    ESP_LOGI(TAG, "wifi_init_sta finished.");
-    EventBits_t bits = xEventGroupWaitBits(s_wifi_event_group,
-                                           WIFI_CONNECTED_BIT | WIFI_FAIL_BIT, pdFALSE, pdFALSE, portMAX_DELAY);
-    if (bits & WIFI_CONNECTED_BIT) {
-        ESP_LOGI(TAG, "connected to ap SSID:%s password:%s",
-                 CONFIG_EXAMPLE_ESP_GMF_WIFI_SSID, CONFIG_EXAMPLE_ESP_GMF_WIFI_PASS);
-    } else if (bits & WIFI_FAIL_BIT) {
-        ESP_LOGI(TAG, "Failed to connect to SSID:%s, password:%s",
-                 CONFIG_EXAMPLE_ESP_GMF_WIFI_SSID, CONFIG_EXAMPLE_ESP_GMF_WIFI_PASS);
-    } else {
-        ESP_LOGE(TAG, "UNEXPECTED EVENT");
-    }
-}
-
-void esp_gmf_teardown_periph_wifi(void)
-{
-    esp_wifi_disconnect();
-    esp_err_t err = esp_wifi_stop();
-    if (err == ESP_ERR_WIFI_NOT_INIT) {
-        ESP_LOGE(TAG, "WIFI stop got an error");
-        return;
-    }
-    ESP_ERROR_CHECK(esp_wifi_deinit());
-    esp_event_loop_delete_default();
-    ESP_ERROR_CHECK(esp_wifi_clear_default_wifi_driver_and_handlers(s_esp_gmf_setup_sta_netif));
-    esp_netif_destroy(s_esp_gmf_setup_sta_netif);
-    s_esp_gmf_setup_sta_netif = NULL;
 }
 
 void esp_gmf_setup_periph_i2c(int port)